+++ demos/agent-task-manager/package.json
@@ -22,6 +22,7 @@
 		"agents-sdk": "^0.0.27",
 		"ai": "^4.1.39",
 		"hono": "^4.7.1",
+		"workers-ai-provider": "0.2.0",
 		"zod": "^3.24.2"
 	}
 }
