+++ nx.json
@@ -18,6 +18,11 @@
 		},
 		"type-check": {
 			"cache": true
+		},
+		"@nx/js:tsc": {
+			"cache": true,
+			"dependsOn": ["^build"],
+			"inputs": ["default", "^default"]
 		}
 	}
 }
