+++ package.json
@@ -2,13 +2,17 @@
 	"name": "ai-mono",
 	"author": "Andy Jessop",
 	"scripts": {
-		"generate-app": "npx tsx ./scripts/generate-app/index.ts",
-		"generate-readmes": "npx tsx ./scripts/readme-generator/readme-generator.ts",
+		"create-demo": "npx tsx ./tools/create-demo/index.ts",
 		"prepare": "husky"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.8.2",
 		"@clack/prompts": "^0.10.0",
+		"@nx/devkit": "20.5.0",
+		"@swc-node/register": "~1.9.1",
+		"@swc/cli": "~0.3.12",
+		"@swc/core": "~1.5.7",
+		"@swc/helpers": "~0.5.11",
 		"@types/ejs": "^3.1.5",
 		"@types/node": "^22.13.10",
 		"@types/wait-on": "^5.3.4",
@@ -18,7 +22,9 @@
 		"husky": "^9.1.7",
 		"miniflare": "^3.20250204.1",
 		"msw": "^2.7.3",
-		"nx": "^20.4.3",
+		"nx": "20.5.0",
+		"prettier": "^2.6.2",
+		"tslib": "^2.3.0",
 		"typescript": "^5.5.2",
 		"vite": "^6.1.0",
 		"vitest": "~3.0.7",
@@ -31,6 +37,7 @@
 		"@changesets/cli": "^2.28.1",
 		"@cloudflare/workers-types": "^4.20250310.0",
 		"@modelcontextprotocol/sdk": "^1.7.0",
+		"@nx/devkit": "20.5.0",
 		"agents-sdk": "^0.0.27",
 		"ai": "^4.1.39",
 		"chalk": "^5.4.1",
@@ -45,9 +52,5 @@
 		"tsup": "^8.4.0",
 		"zod": "^3.24.2"
 	},
-	"workspaces": [
-		"libs/*",
-		"demos/*",
-		"packages/*"
-	]
+	"workspaces": ["libs/*", "demos/*", "packages/*"]
 }
