--- scripts/generate-app/servers/structured-output/index.ts.ejs
@@ -1,49 +1,0 @@
-import { Hono } from "hono";
-import { cors } from "hono/cors";
-import { generateObject } from "ai";
-import z from "zod";
-import type { Env } from "./types/env.ts";
-import type { Variables } from "./types/hono.ts";
-<% if (provider === 'openai') { -%>
-import { createOpenAI } from "@ai-sdk/openai";
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-import { createWorkersAI } from "../../../packages/workers-ai-provider/src";
-<% } -%>
-
-const app = new Hono<{ Bindings: Env; Variables: Variables }>();
-app.use(cors());
-
-app.post("/", async (c) => {
-	const { prompt } = (await c.req.json()) as { prompt: string };
-<% if (provider === 'openai') { -%>
-    const openai = createOpenAI({
-		apiKey: c.env.OPENAI_API_KEY,
-	});
-	const model = openai("gpt-4o-mini");
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-    const workersai = createWorkersAI({ binding: c.env.AI });
-    const model = workersai("@cf/meta/llama-3.3-70b-instruct-fp8-fast");
-<% } -%>
-
-	const { object } = await generateObject({
-		model,
-		schema: z.object({
-			recipe: z.object({
-				name: z.string(),
-				ingredients: z.array(
-					z.object({ name: z.string(), amount: z.string() }),
-				),
-				steps: z.array(z.string()),
-			}),
-		}),
-		prompt,
-	});
-
-	return c.json(object);
-});
-
-export default {
-	fetch: app.fetch,
-} satisfies ExportedHandler<Env>;
