--- scripts/generate-app/servers/text-generation/index.ts.ejs
@@ -1,39 +1,0 @@
-import { Hono } from "hono";
-import { cors } from "hono/cors";
-import { generateText } from "ai";
-<% if (provider === 'openai') { -%>
-import { createOpenAI } from "@ai-sdk/openai";
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-import { createWorkersAI } from "../../../packages/workers-ai-provider/src";
-<% } -%>
-import type { Env } from "./types/env.ts";
-import type { Variables } from "./types/hono.ts";
-
-const app = new Hono<{ Bindings: Env; Variables: Variables }>();
-app.use(cors());
-
-app.post("/", async (c) => {
-	const { prompt } = (await c.req.json()) as { prompt: string };
-<% if (provider === 'openai') { -%>
-    const openai = createOpenAI({
-		apiKey: c.env.OPENAI_API_KEY,
-	});
-	const model = openai("gpt-4o-mini");
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-    const workersai = createWorkersAI({ binding: c.env.AI });
-    const model = workersai("@cf/meta/llama-3.3-70b-instruct-fp8-fast");
-<% } -%>
-
-	const result = await generateText({
-		model,
-		prompt,
-	});
-
-	return c.json(result);
-});
-
-export default {
-	fetch: app.fetch,
-} satisfies ExportedHandler<Env>;
