--- scripts/generate-app/servers/tool-calling/index.ts.ejs
@@ -1,56 +1,0 @@
-import { Hono } from "hono";
-import { cors } from "hono/cors";
-import { generateText, tool } from "ai";
-import z from "zod";
-<% if (provider === 'openai') { -%>
-import { createOpenAI } from "@ai-sdk/openai";
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-import { createWorkersAI } from "../../../packages/workers-ai-provider/src";
-<% } -%>
-import type { Env } from "./types/env.ts";
-import type { Variables } from "./types/hono.ts";
-
-const app = new Hono<{ Bindings: Env; Variables: Variables }>();
-app.use(cors());
-
-app.post("/", async (c) => {
-	const { prompt } = (await c.req.json()) as { prompt: string };
-<% if (provider === 'openai') { -%>
-    const openai = createOpenAI({
-		apiKey: c.env.OPENAI_API_KEY,
-	});
-	const model = openai("gpt-4o-mini");
-<% } -%>
-<% if (provider === 'workers-ai') { -%>
-    const workersai = createWorkersAI({ binding: c.env.AI });
-    const model = workersai("@cf/meta/llama-3.3-70b-instruct-fp8-fast");
-<% } -%>
-
-	const result = await generateText({
-		model,
-		messages: [
-			{ role: "system", content: "You are a helpful AI assistant" },
-			{ role: "user", content: prompt },
-		],
-		tools: {
-			weather: tool({
-				description: "Get the weather in a location",
-				parameters: z.object({
-					location: z.string().describe("The location to get the weather for"),
-				}),
-				execute: async ({ location }) => ({
-					location,
-					weather: location === "London" ? "Raining" : "Sunny",
-				}),
-			}),
-		},
-        maxSteps: 5,
-	});
-
-	return c.json(result);
-});
-
-export default {
-	fetch: app.fetch,
-} satisfies ExportedHandler<Env>;
