--- scripts/generate-app/worker-shell-with-ui/package.json.ejs
@@ -1,14 +1,0 @@
-{
-	"name": "<%= projectName -%>",
-	"type": "module",
-	"scripts": {
-		"dev": "vite",
-		"build": "vite build",
-		"lint": "biome lint --error-on-warnings ./src",
-		"preview": "npm run build && vite preview",
-		"deploy": "npm run build && wrangler deploy",
-		"test": "vitest --watch=false",
-		"test:ci": "vitest --watch=false",
-		"type-check": "tsc --noEmit"
-	}
-}
