--- scripts/generate-app/worker-shell-with-ui/src/client/App.tsx.ejs
@@ -1,5 +1,0 @@
-function App() {
-	return <>App</>;
-}
-
-export default App;
