--- scripts/generate-app/worker-shell-with-ui/src/client/main.tsx.ejs
@@ -1,16 +1,0 @@
-import { MantineProvider, createTheme } from "@mantine/core";
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-import App from "./App";
-
-const theme = createTheme({
-	/** Put your mantine theme override here */
-});
-
-createRoot(document.getElementById("root")!).render(
-	<StrictMode>
-		<MantineProvider theme={theme} withNormalizeCSS withGlobalStyles>
-			<App />
-		</MantineProvider>
-	</StrictMode>,
-);
