--- scripts/generate-app/worker-shell-with-ui/src/server/types/env.ts.ejs
@@ -1,4 +1,0 @@
-export interface Env {
-	ENVIRONMENT: "production" | "development" | "staging";
-	ASSETS: Fetcher;
-}
