--- scripts/generate-app/worker-shell-with-ui/wrangler.jsonc.ejs
@@ -1,19 +1,0 @@
-{
-	"$schema": "../../node_modules/wrangler/config-schema.json",
-	"name": "<%= projectName -%>",
-	"main": "src/server/index.ts",
-	"compatibility_date": "2025-02-24",
-	"assets": {
-		"not_found_handling": "single-page-application",
-		"binding": "ASSETS",
-        "directory": "./dist/client"
-	},
-	"observability": {
-		"enabled": true
-	}
-    <% if (provider === 'workers-ai') { %>,
-    "ai": {
-        "binding": "AI"
-    }
-    <% } -%>
-}
