--- scripts/generate-app/worker-shell/.dev.vars.ejs
@@ -1,3 +1,0 @@
-<% if (provider === 'openai') { -%>
-OPENAI_API_KEY=
-<% } -%>
