--- scripts/generate-app/worker-shell/package.json.ejs
@@ -1,12 +1,0 @@
-{
-	"name": "<%= projectName -%>",
-	"scripts": {
-		"deploy": "wrangler deploy",
-		"dev": "wrangler dev",
-		"lint": "biome lint --error-on-warnings ./src",
-		"start": "wrangler",
-		"test": "vitest --watch=false",
-		"test:ci": "vitest --watch=false",
-		"type-check": "tsc --noEmit"
-	}
-}
