--- scripts/generate-app/worker-shell/src/types/env.ts.ejs
@@ -1,7 +1,0 @@
-export interface Env {
-	ENVIRONMENT: "production" | "development" | "staging";
-<% if (provider === 'openai') { -%>
-    OPENAI_API_KEY: string;
-<% } -%>
-	AI: Ai;
-}
