--- scripts/generate-app/worker-shell/src/types/hono.ts.ejs
@@ -1,17 +1,0 @@
-import type {
-	Hono,
-	Context as HonoContext,
-	MiddlewareHandler as HonoMiddlewareHandler,
-} from "hono";
-import type { Env } from "./env";
-
-export type Variables = Record<string, string>;
-
-export type App = Hono<{ Bindings: Env; Variables: Variables }>;
-
-export type MiddlewareHandler = HonoMiddlewareHandler<{
-	Bindings: Env;
-	Variables: Variables;
-}>;
-export type Context = HonoContext<{ Bindings: Env; Variables: Variables }>;
-export type Next = () => Promise<void>;
