--- scripts/generate-app/worker-shell/wrangler.jsonc.ejs
@@ -1,14 +1,0 @@
-{
-	"$schema": "../../node_modules/wrangler/config-schema.json",
-	"name": "<%= projectName -%>",
-	"main": "src/index.ts",
-	"compatibility_date": "2025-02-24",
-	"observability": {
-		"enabled": true
-	}
-    <% if (provider === 'workers-ai') { %>,
-    "ai": {
-        "binding": "AI"
-    }
-    <% } -%>
-}
