+++ tools/create-demo/scaffolding/worker-with-client/README.md.ejs
@@ -1,0 +1,1 @@
+# <%= projectName -%>
