+++ tools/create-demo/scaffolding/worker-with-client/src/client/App.tsx.ejs
@@ -1,0 +1,5 @@
+function App() {
+	return <><%= projectName -%></>;
+}
+
+export default App;
