+++ tools/create-demo/scaffolding/worker-with-client/src/client/main.tsx.ejs
@@ -1,0 +1,9 @@
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import App from "./App";
+
+createRoot(document.getElementById("root")!).render(
+	<StrictMode>
+        <App />
+	</StrictMode>,
+);
