+++ tools/create-demo/scaffolding/worker-with-client/src/client/vite-env.d.ts.ejs
