+++ tools/create-demo/scaffolding/worker-with-client/src/server/types/hono.ts.ejs
@@ -1,0 +1,16 @@
+import type {
+	Hono,
+	Context as HonoContext,
+	MiddlewareHandler as HonoMiddlewareHandler,
+} from "hono";
+
+export type Variables = Record<string, string>;
+
+export type App = Hono<{ Bindings: Env; Variables: Variables }>;
+
+export type MiddlewareHandler = HonoMiddlewareHandler<{
+	Bindings: Env;
+	Variables: Variables;
+}>;
+export type Context = HonoContext<{ Bindings: Env; Variables: Variables }>;
+export type Next = () => Promise<void>;
