+++ tools/create-demo/scaffolding/worker-with-client/test/tsconfig.json.ejs
@@ -1,0 +1,7 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"types": ["@cloudflare/vitest-pool-workers"]
+	},
+	"include": ["../worker-configuration.d.ts", "**/*.ts"]
+}
