+++ tools/create-demo/scaffolding/worker-with-client/vite.config.ts.ejs
