+++ tools/create-demo/scaffolding/worker-with-client/vitest.config.ts.ejs
