+++ tools/create-demo/scaffolding/worker-with-client/wrangler.jsonc.ejs
@@ -1,0 +1,17 @@
+{
+	"$schema": "node_modules/wrangler/config-schema.json",
+	"name": "<%= projectName -%>",
+	"main": "src/server/index.ts",
+	"compatibility_date": "2025-02-24",
+    "assets": {
+		"not_found_handling": "single-page-application",
+		"binding": "ASSETS",
+        "directory": "./dist/client"
+	},
+	"observability": {
+		"enabled": true
+	},
+    "ai": {
+        "binding": "AI"
+    }
+}
