+++ tools/create-demo/scaffolding/worker/README.md.ejs
@@ -1,0 +1,1 @@
+# <%= projectName -%>
