+++ tools/create-demo/scaffolding/worker/src/index.ts.ejs
@@ -1,0 +1,12 @@
+import { Hono } from "hono";
+import { cors } from "hono/cors";
+import type { Variables } from "./types/hono";
+
+const app = new Hono<{ Bindings: Env; Variables: Variables }>();
+
+app.use(cors());
+app.get("/", (c) => c.json("ok"));
+
+export default {
+	fetch: app.fetch,
+} satisfies ExportedHandler<Env>;
