+++ tools/create-demo/scaffolding/worker/vite.config.ts.ejs
@@ -1,0 +1,6 @@
+import { cloudflare } from "@cloudflare/vite-plugin";
+import { defineConfig } from "vite";
+
+export default defineConfig({
+	plugins: [cloudflare()],
+});
