+++ tools/create-demo/scaffolding/worker/vitest.config.ts.ejs
