+++ tools/create-demo/scaffolding/worker/wrangler.jsonc.ejs
@@ -1,0 +1,12 @@
+{
+	"$schema": "node_modules/wrangler/config-schema.json",
+	"name": "<%= projectName -%>",
+	"main": "src/index.ts",
+	"compatibility_date": "2025-02-24",
+	"observability": {
+		"enabled": true
+	},
+    "ai": {
+        "binding": "AI"
+    }
+}
