+++ tsconfig.node.json
@@ -1,7 +1,9 @@
 {
 	"extends": "./tsconfig.base.json",
-	"include": ["./**/*.ts",
-		"./demos/structured-output-node/src/**/*.ts"
+	"include": [
+		"./**/*.ts",
+		"./demos/structured-output-node/src/**/*.ts",
+		"./tools/create-demo/src/**/*.ts"
 	],
 	"exclude": ["**/src/**/*.ts", "**/test/**/*.ts"]
 }
